requirements.txt
asgiref==3.9.1
django==5.2.6
pillow==11.3.0
sqlparse==0.5.3
